Update 2023aug23addendum.md
Add total column.
Note that some of the numbers have slightly changed, for the cases where the bonus is added without adjusting taxes.
When the deficits change, I think it's because I'm using the actual deficits instead of the shortages???
No idea why the other numbers very slightly change in the 0.5 case, though.