Update supplyanddemand.md
Removed the kg stuff and set the background color to solarized.
I'd still like to embed graphs in the slide, but I think I'm going to have to admit that remark.js is just too jank to use.

I should stick with lyx slides or something, and have seperately hosted pages for interactive js stuff I want to include.